+++ src/app/error-boundary/error-boundary.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { Component } from 'react';
 
-import ErrorBoundary from '../ErrorBoundary';
+import ErrorBoundary from './error-boundary';
 
 class ProblematicComponent extends Component {
   render() {
