+++ src/components/description/description.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useRef, useState } from 'react';
 import { Link, useNavigate, useParams } from 'react-router';
 
-import { LINK } from '../constants';
-import { usePagination } from '../hooks';
-import { Loader } from './loader';
+import { Loader } from '../';
+import { LINK } from '../../constants';
+import { usePagination } from '../../hooks';
 
 interface AbilityItem {
   ability: {
