+++ src/components/flyout/flyout.test.tsx
@@ -2,13 +2,10 @@ import { render, screen, fireEvent } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 
-import { Flyout } from '../components/flyout';
-import {
-  clearSelected,
-  downloadSelected,
-} from '../components/slices/selected-slice';
+import { Flyout } from './flyout';
+import { clearSelected, downloadSelected } from '../../features/selected-slice';
 
-jest.mock('../components/slices/selected-slice', () => ({
+jest.mock('../../features/selected-slice', () => ({
   clearSelected: jest.fn(() => ({ type: 'CLEAR_SELECTED' })),
   downloadSelected: jest.fn(() => ({ type: 'DOWNLOAD_SELECTED' })),
 }));
