+++ src/components/index.tsx
@@ -1,11 +1,12 @@
-export { Card } from './card';
-export { CardList } from './card-list';
-export { Loader } from './loader';
-export { Search } from './search';
-export { Description } from './description';
-export { Home } from './home';
+export { Card } from './card/card';
+export { CardList } from './card-list/card-list';
+export { Loader } from './loader/loader';
+export { Search } from './search/search';
+export { Description } from './description/description';
+export { Home } from './home/home';
 export { About } from '../pages/about';
-export { Header } from './header';
+export { Header } from './header/header';
 export { NotFound } from '../pages/404';
-export { Selected } from './selected';
-export { Flyout } from './flyout';
+export { Selected } from './selected/selected';
+export { Flyout } from './flyout/flyout';
+export { ThemeToggle } from './theme-toggle/theme-toggle';
