+++ src/components/search/search.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent } from '@testing-library/react';
 
-import { Search } from '../components/search';
+import { Search } from '../';
 
 describe('Search Component', () => {
   beforeEach(() => {
