+++ src/features/selected-slice.ts
@@ -1,8 +1,8 @@
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
 
-import { downloadSelectedItems } from '../../utils/download-selected-items';
+import { downloadSelectedItems } from '../utils/download-selected-items';
 
-import type { Pokemons } from '../../types';
+import type { Pokemons } from '../types';
 
 const selectedSlice = createSlice({
   name: 'selectedItems',
