+++ src/hooks/index.tsx
@@ -1,2 +1,2 @@
-export { useLocalStorage } from './useLocalStorage';
-export { usePagination } from './usePagination';
+export { useLocalStorage } from './use-local-storage';
+export { usePagination } from './use-pagination';
