+++ src/main.tsx
@@ -4,10 +4,10 @@ import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router';
 
-import App from './App';
-import { store } from './components/store';
-import { ThemeProvider } from './components/theme-context';
-import ErrorBoundary from './ErrorBoundary';
+import App from './app/App';
+import ErrorBoundary from './app/error-boundary/error-boundary';
+import { store } from './app/store';
+import { ThemeProvider } from './features/theme-context';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
